requirements.txt
pandas
openpyxl
pymongo
snakecase
requests
google-cloud-tasks==2.10.4
google-cloud-pubsub==2.8.0
flask[async]
Flask==2.1.0
gunicorn==20.1.0